refactor(accountProxy): update import paths for hooks and components
